changes to restart
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@252 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/shell8/s8_main.c
@@ -19,7 +19,7 @@ void shell8(FIELD      *actfield,
             ARRAY      *intforce_global,
             int         kstep,
             int         handsize,
-            int        *handles,
+            long int   *handles,
             CALC_ACTION *action)
 {
 #ifdef D_SHELL8
